Rename watchPrice saga
