+++ src/state/saga.js
@@ -10,7 +10,7 @@ import {
 import { sendStartSaga } from './sagas/send';
 import submitTransactionSaga from './sagas/submitTransaction';
 import watchAccountSaga from './sagas/watchAccount';
-import watchPricesSaga from './sagas/watchPricesSaga';
+import watchPricesSaga from './sagas/watchCoinMarketCapSaga';
 import { sendHackathonVoteSaga } from './sagas/hackathonVote';
 import watchPaymentsSaga from './sagas/watchPayments';
